test_drivers.py
"""
Test database driver imports and basic connectivity
"""

print("Testing Database Drivers Installation\n")
print("=" * 60)

# Test 1: PostgreSQL Driver
print("\n1. PostgreSQL Driver (psycopg2-binary):")
try:
    import psycopg2
    print(f"   ✓ psycopg2 version: {psycopg2.__version__}")
    print("   ✓ Ready to connect to PostgreSQL databases")
except ImportError as e:
    print(f"   ✗ Failed: {e}")

# Test 2: MySQL Driver
print("\n2. MySQL Driver (pymysql):")
try:
    import pymysql
    print(f"   ✓ pymysql version: {pymysql.__version__}")
    print("   ✓ Ready to connect to MySQL/MariaDB databases")
except ImportError as e:
    print(f"   ✗ Failed: {e}")

# Test 3: SQL Server Driver
print("\n3. SQL Server Driver (pyodbc):")
try:
    import pyodbc
    print(f"   ✓ pyodbc version: {pyodbc.version}")
    print("   ✓ Available ODBC drivers:")
    drivers = pyodbc.drivers()
    if drivers:
        for driver in drivers[:5]:  # Show first 5
            print(f"      - {driver}")
        if len(drivers) > 5:
            print(f"      ... and {len(drivers) - 5} more")
    else:
        print("      (No ODBC drivers found - install SQL Server ODBC driver)")
    print("   ✓ Ready to connect to SQL Server databases")
except ImportError as e:
    print(f"   ✗ Failed: {e}")

# Test 4: SQLite (built-in)
print("\n4. SQLite Driver (built-in):")
try:
    import sqlite3
    print(f"   ✓ sqlite3 version: {sqlite3.sqlite_version}")
    print("   ✓ Ready to connect to SQLite databases")
except ImportError as e:
    print(f"   ✗ Failed: {e}")

# Test 5: SQLAlchemy connection strings
print("\n" + "=" * 60)
print("\nSupported Connection String Examples:")
print("\nPostgreSQL:")
print("  postgresql://user:password@localhost:5432/database")
print("  postgresql+psycopg2://user:password@localhost:5432/database")

print("\nMySQL:")
print("  mysql+pymysql://user:password@localhost:3306/database")

print("\nSQL Server:")
print("  mssql+pyodbc://user:password@localhost:1433/database?driver=ODBC+Driver+17+for+SQL+Server")
print("  mssql+pyodbc://user:password@localhost/database?driver=SQL+Server")

print("\nSQLite:")
print("  sqlite:///path/to/database.db")
print("  sqlite:///C:/path/to/database.db  (Windows)")

print("\n" + "=" * 60)
print("\n✓ All database drivers installed successfully!")
print("\nYou can now use DatabaseValidator with:")
print("  - PostgreSQL databases")
print("  - MySQL/MariaDB databases")
print("  - SQL Server databases")
print("  - SQLite databases")
print("\n" + "=" * 60)
